correction de la sauvegarde de l'utilisateur du login des roles

+++ src/main/java/com/dsf/escalade/repository/global/UtilisateurRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UtilisateurRepository extends JpaRepository<Utilisateur,Integer> {
+public interface UtilisateurRepository extends JpaRepository<Utilisateur,Long> {
    Utilisateur findByNom(String nom);
    Utilisateur findByMail(String email);
    Utilisateur findByLogin(String login);
